requirements.txt
beautifulsoup4
boto3
line-bot-sdk
awscli
